Added group migration on login & email Intial setup

+++ src/lib/userGroupsHelper.ts
@@ -22,6 +22,7 @@ async function getSessionEmail() {
   
 async function getUserIdByEmail(): Promise<string | null> {
     const email = await getSessionEmail() as string
+    if (!email) return null
     const user = await prisma.user.findUnique({
       where: { email },
       select: { id: true },
@@ -31,6 +32,7 @@ async function getUserIdByEmail(): Promise<string | null> {
 
   export async function getRecentGroupsDB(): Promise<RecentGroups> {
     const userId = await getUserIdByEmail() as string
+    if (!userId) return []
     const recentGroups = await prisma.recentGroup.findMany({
       where: { userId },
       orderBy: { id: 'desc' },
@@ -40,6 +42,7 @@ async function getUserIdByEmail(): Promise<string | null> {
   
   export async function saveRecentGroupDB(group: RecentGroup) {
     const userId = await getUserIdByEmail() as string
+    if (!userId) return []
     await prisma.recentGroup.upsert({
       where: { userId_groupId: { groupId: group.id, userId } },
       update: { name: group.name },
@@ -49,6 +52,7 @@ async function getUserIdByEmail(): Promise<string | null> {
   
   export async function deleteRecentGroupDB(groupId: string) {
     const userId = await getUserIdByEmail() as string
+    if (!userId) return []
     await prisma.recentGroup.deleteMany({
       where: { groupId, userId },
     })
@@ -56,6 +60,7 @@ async function getUserIdByEmail(): Promise<string | null> {
   
   export async function getStarredGroupsDB(): Promise<z.infer<typeof starredGroupsSchema>> {
     const userId = await getUserIdByEmail() as string
+    if (!userId) return []
     const starredGroups = await prisma.starredGroup.findMany({
       where: { userId },
     })
@@ -64,6 +69,7 @@ async function getUserIdByEmail(): Promise<string | null> {
   
   export async function starGroupDB(groupId: string) {
     const userId = await getUserIdByEmail() as string
+    if (!userId) return []
     await prisma.starredGroup.create({
       data: { groupId, userId },
     })
@@ -71,6 +77,7 @@ async function getUserIdByEmail(): Promise<string | null> {
   
   export async function unstarGroupDB(groupId: string) {
     const userId = await getUserIdByEmail() as string
+    if (!userId) return []
     await prisma.starredGroup.deleteMany({
       where: { groupId, userId },
     })
@@ -78,6 +85,7 @@ async function getUserIdByEmail(): Promise<string | null> {
   
   export async function getArchivedGroupsDB(): Promise<z.infer<typeof archivedGroupsSchema>> {
     const userId = await getUserIdByEmail() as string
+    if (!userId) return []
     const archivedGroups = await prisma.archivedGroup.findMany({
       where: { userId },
     })
@@ -86,6 +94,7 @@ async function getUserIdByEmail(): Promise<string | null> {
   
   export async function archiveGroupDB(groupId: string) {
     const userId = await getUserIdByEmail() as string
+    if (!userId) return []
     await prisma.archivedGroup.create({
       data: { groupId, userId },
     })
@@ -93,7 +102,32 @@ async function getUserIdByEmail(): Promise<string | null> {
   
   export async function unarchiveGroupDB(groupId: string) {
     const userId = await getUserIdByEmail() as string
+    if (!userId) return []
     await prisma.archivedGroup.deleteMany({
       where: { groupId, userId },
     })
+  }
+
+  export async function getEmailsByGroupId(groupId: string): Promise<{ email: string; name: string }[]> {
+    if (!groupId) return []
+    // Fetch all users associated with the given groupId from the RecentGroup model
+    const users = await prisma.recentGroup.findMany({
+      where: {
+        groupId: groupId,
+      },
+      select: {
+        user: {
+          select: {
+            email: true,
+            name: true
+          },
+        },
+      },
+    });
+  
+    // Map the result to return an array of emails
+    return users.map((group) => ({
+      email: group.user.email,
+      name: group.user.name,
+    }));
   }
